Project_1_RA/requirements.txt
google-genai
googlesearch-python
